refactor: post 훅들 comment와 분리
