+++ src/entities/post/api/index.ts
@@ -16,10 +16,8 @@ export const fetchTags = async () => {
   return data
 }
 
-export const fetchPosts = async ({ limit, skip, searchQuery, tag, sortBy, sortOrder }: SearchParams) => {
-  const response = await fetch(
-    `/api/posts/search?q=${searchQuery}&limit=${limit}&skip=${skip}&tag=${tag}&sortBy=${sortBy}&order=${sortOrder}`,
-  )
+export const fetchPosts = async ({ limit, skip, sortBy, sortOrder }: SearchParams) => {
+  const response = await fetch(`/api/posts?limit=${limit}&skip=${skip}&sortBy=${sortBy}&order=${sortOrder}`)
   const data: PostResponse = await response.json()
   return data
 }
