+++ src/features/comment/model/index.ts
@@ -1,0 +1,5 @@
+export * from "./useCommentsQuery.ts"
+export * from "./useLikeCommentsMutation.ts"
+export * from "./useUpdateCommentMutation.ts"
+export * from "./useAddCommentMutation.ts"
+export * from "./useDeleteCommentMutation.ts"
