+++ src/features/post/lib/index.ts
@@ -1,4 +1,4 @@
-import { fetchPosts } from "@/entities/post/api"
+import { fetchPosts, searchPost, searchPostByTag } from "@/entities/post/api"
 import { fetchUsers } from "@/entities/user/api"
 import { PostResponse, PostWithUser, SearchParams } from "@entities/post/types"
 import { User } from "@entities/user/types"
@@ -17,12 +17,12 @@ export const getPostsWithUsers = async (params: SearchParams) => {
   return combinePostsWithUser(postsData, usersData)
 }
 
-// export const getPostsWithUserByTag = async (tag: string) => {
-//   const [postsData, usersData] = await Promise.all([searchPostByTag(tag), fetchUsers()])
-//   return combinePostsWithUser(postsData, usersData)
-// }
-//
-// export const getPostsWithUserBySearch = async (searchQuery: string) => {
-//   const [postsData, usersData] = await Promise.all([searchPost(searchQuery), fetchUsers()])
-//   return combinePostsWithUser(postsData, usersData)
-// }
+export const getPostsWithUserByTag = async (tag: string) => {
+  const [postsData, usersData] = await Promise.all([searchPostByTag(tag), fetchUsers()])
+  return combinePostsWithUser(postsData, usersData)
+}
+
+export const getPostsWithUserBySearch = async (searchQuery: string) => {
+  const [postsData, usersData] = await Promise.all([searchPost(searchQuery), fetchUsers()])
+  return combinePostsWithUser(postsData, usersData)
+}
