+++ src/features/post/ui/CommentAdd.tsx
@@ -1,6 +1,7 @@
 import { Button, DialogContent, DialogHeader, DialogTitle, Textarea } from "@shared/ui"
 import { useState } from "react"
-import { useAddCommentMutation } from "@features/post/model"
+import { useAddCommentMutation } from "@features/comment/model"
+import { useModalStore } from "@shared/model"
 
 interface CommentAddProps {
   postId: number
@@ -10,9 +11,11 @@ export function CommentAdd(props: CommentAddProps) {
   const { postId } = props
   const [newComment, setNewComment] = useState({ body: "", postId: postId, userId: 1 })
   const { mutate: addComment } = useAddCommentMutation()
+  const { closeModal } = useModalStore()
 
   const handleAddComment = () => {
     addComment(newComment)
+    closeModal()
   }
 
   return (
